requirements-production.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
langchain-google-genai==0.0.6
langchain==0.1.4
langchain_community==0.0.16
chromadb==0.4.22
sentence-transformers==2.3.1
scikit-learn==1.4.0
python-dotenv==1.0.0
slowapi==0.1.9
pydantic==2.5.3
html5lib==1.1
